Restructured tests, currently it doesn't work
