+++ ask/transpiler/tests/test_errors.py
@@ -1,0 +1,14 @@
+import unittest
+
+from ask.transpiler import errors
+
+
+class TestAskTranspilerErrors(unittest.TestCase):
+
+
+def main():
+	unittest.main()
+
+
+if __name__ == '__main__':
+	main()
