+++ ask/transpiler/tests/test_parser.py
@@ -1,0 +1,14 @@
+import unittest
+
+from ask.transpiler import parser
+
+
+class TestAskTranspilerParser(unittest.TestCase):
+
+
+def main():
+	unittest.main()
+
+
+if __name__ == '__main__':
+	main()
