+++ ask/transpiler/tests/test_transpiler.py
@@ -1,0 +1,14 @@
+import unittest
+
+from ask.transpiler import transpiler
+
+
+class TestAskTranspilerTranspiler(unittest.TestCase):
+
+
+def main():
+	unittest.main()
+
+
+if __name__ == '__main__':
+	main()
