+++ ask/transpiler/utilities/tests/test_lexer_utils.py
@@ -1,0 +1,111 @@
+import unittest
+
+from ask.transpiler.utilities import lexer_utils
+
+
+class TestAskTranspilerUtilitiesLexerUtils(unittest.TestCase):
+	# tokens_grouped_by_lines()
+	def test_tokens_grouped_by_lines(self):
+		tokens = [
+			['VAR', 'variable'],
+			['FORMAT', '\n'],
+			['VAR', 'variable'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['VAR', 'variable'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['FORMAT', '\t'],
+			['VAR', 'variable'],
+		]
+
+		expected = [
+			[
+				['VAR', 'variable'],
+			],
+			[
+				['FORMAT', '\n'],
+				['VAR', 'variable'],
+			],
+			[
+				['FORMAT', '\n'],
+				['FORMAT', '\t'],
+				['VAR', 'variable'],
+			],
+			[
+				['FORMAT', '\n'],
+				['FORMAT', '\t'],
+				['FORMAT', '\t'],
+				['VAR', 'variable'],
+			],
+		]
+
+		self.assertEqual(expected, lexer_utils.tokens_grouped_by_lines(tokens))
+
+	# insert_indentation_group_markers()
+	def test_insert_indention_group_markers(self):
+		tokens = [
+			['FUNC_DEF', 'function'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['FUNC', 'print'],
+			['STR', 'text'],
+			['OP', ')'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['KEYWORD', 'if'],
+			['VAR', 'var'],
+			['OP', ':'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['FORMAT', '\t'],
+			['FUNC', 'print'],
+			['STR', 'text'],
+			['OP', ')'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['FUNC', 'print'],
+			['STR', 'text'],
+			['OP', ')'],
+		]
+
+		expected = [
+			['GROUP', 'start'],
+			['FUNC_DEF', 'function'],
+			['GROUP', 'start'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['FUNC', 'print'],
+			['STR', 'text'],
+			['OP', ')'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['KEYWORD', 'if'],
+			['VAR', 'var'],
+			['OP', ':'],
+			['GROUP', 'start'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['FORMAT', '\t'],
+			['FUNC', 'print'],
+			['STR', 'text'],
+			['OP', ')'],
+			['GROUP', 'end'],
+			['FORMAT', '\n'],
+			['FORMAT', '\t'],
+			['FUNC', 'print'],
+			['STR', 'text'],
+			['OP', ')'],
+			['GROUP', 'end'],
+			['GROUP', 'end']
+		]
+
+		self.assertEqual(expected, lexer_utils.insert_indention_group_markers(tokens))
+
+
+def main():
+	unittest.main()
+
+
+if __name__ == '__main__':
+	main()
