+++ ask/transpiler/utilities/tests/test_small_transpilers.py
@@ -1,0 +1,14 @@
+import unittest
+
+from ask.transpiler.utilities import small_transpilers
+
+
+class TestAskTranspilerUtilitiesSmallTranspilers(unittest.TestCase):
+
+
+def main():
+	unittest.main()
+
+
+if __name__ == '__main__':
+	main()
