+++ ask/transpiler/utilities/tests/test_transpiler_utils.py
@@ -1,0 +1,14 @@
+import unittest
+
+from ask.transpiler.utilities import transpiler_utils
+
+
+class TestAskTranspilerUtilitiesTranspilerUtils(unittest.TestCase):
+
+
+def main():
+	unittest.main()
+
+
+if __name__ == '__main__':
+	main()
