+++ ask/utilities/tests/test_utils.py
@@ -1,0 +1,14 @@
+import unittest
+
+from ask.utilities import utils
+
+
+class TestAskUtilitiesUtils(unittest.TestCase):
+
+
+def main():
+	unittest.main()
+
+
+if __name__ == '__main__':
+	main()
